LIVE-3021: move interview kicker to headerMedia component
